requirements.txt
aiofiles
aiohttp
aiosqlite
discord.py
python-dotenv